Move Servers page to pages directory
