+++ dashboard/src/pages/styles/Server.tsx
